+++ components/market/myLink/reviews.tsx
@@ -13,7 +13,10 @@ const Reviews = ({ data }: { data: IReviews | null }) => {
     <>
       {data && data.reviews.length > 0 && (
         <div key={"reviews"} id="reviews" className={styles.all}>
-          <div className={styles.header} onClick={toggleContentVisibility}>
+          <div
+            className={styles.header}
+            // onClick={toggleContentVisibility}
+          >
             <div className={`${styles.squre} ${!isContentVisible ? styles.closed : ""}`}></div>
             <div className={styles.headertext}>
               Reviews